Implementación de validaciones, y mejora en formulario

+++ app/Http/Controllers/ArticuloController.php
@@ -23,13 +23,32 @@ class ArticuloController extends Controller
 
     public function store(Request $request)
     {
+        $campos=[
+            'Nombre'=>'required|string|max:100',
+             'Marca'=>'required|string|max:100',
+             'Modelo'=>'required|string|max:100',
+             'NumSerie'=>'required|string|max:100',
+             'Estado'=>'required|string|max:100',
+             'Ubicacion'=>'required|string|max:100',
+             'Foto'=>'required|max:10000|mimes:jpeg,png,jpg',
+        ];
+
+        $mensaje=[
+            'required'=> 'El :attribute del artículo es obligatorio',
+            'Marca.required'=>'La :attribute del artículo es obligatoria',
+            'Ubicacion.required'=>'La :attribute del artículo es obligatoria',
+            'Foto.required'=>'La foto es obligatoria',
+        ];
+
+        $this->validate($request, $campos,$mensaje);
+
         $datos = request()->except(['_token']);
         if ($request->hasFile('Foto')) {
             $datos['Foto'] = $request->file('Foto')->store('uploads', 'public');
         }
 
         Articulo::create($datos);
-        return redirect('articulo')->with('mensaje', 'Articulo agregado');
+        return redirect('articulo')->with('mensaje', 'Artículo agregado con éxito');
     }
 
     public function show($id)
@@ -46,6 +65,29 @@ class ArticuloController extends Controller
 
     public function update(Request $request, $id)
     {
+        $campos=[
+            'Nombre'=>'required|string|max:100',
+             'Marca'=>'required|string|max:100',
+             'Modelo'=>'required|string|max:100',
+             'NumSerie'=>'required|string|max:100',
+             'Estado'=>'required|string|max:100',
+             'Ubicacion'=>'required|string|max:100',
+        ];
+
+        $mensaje=[
+            'required'=> 'El :attribute del artículo es obligatorio',
+            'Marca.required'=>'La :attribute del artículo es obligatoria',
+            'Ubicacion.required'=>'La :attribute del artículo es obligatoria',
+            'Foto.required'=>'La foto es obligatoria',
+        ];
+
+        if ($request->hasFile('Foto')){
+            $campos=['Foto'=>'required|max:10000|mimes:jpeg,png,jpg',];
+            $mensaje=['Foto.required'=>'La foto es obligatoria',];
+        } 
+
+        $this->validate($request, $campos,$mensaje);
+
         $datos = request()->except(['_token', '_method']);
         
         if ($request->hasFile('Foto')) {
@@ -62,7 +104,12 @@ class ArticuloController extends Controller
 
     public function destroy($id)
     {
-        Articulo::destroy($id);
-        return redirect('articulo');
+        $articulo = Articulo::findOrFail($id);
+
+        if(Storage::delete('public/'.$articulo->Foto)){
+            Articulo::destroy($id);
+        }
+        
+        return redirect('articulo')->with('mensaje', 'Artículo eliminado con éxito');
     }
 }
